Extract calibration-form population UI into engine/src/depth/calibrationForm.ts
Moves displayCalibrationInfo, onCameraSelectionChange, and
populateFormFromCalibration out of PointCloudVisualizer - populating
the depth-settings form (fx/fy/cx/cy, baseline, distortion
coefficients, camera model) from a parsed calibration file, including
the baseline-present auto-switch to disparity mode.

main.ts: 11,437 -> 11,225 lines.

Verified: engine tsc clean, full webpack compile clean, all 59
extension tests passing under xvfb-run.
